Move BibleBook into own interface & rename folder to models
