+++ src/bible/bible-api/bible-api.class.ts
@@ -1,5 +1,5 @@
 import { Bible } from 'src/bible/bible.class';
-import { BibleVersionCode } from 'src/bible/versions/bible-versions.const';
+import { BibleVersionCode } from 'src/bible/models/bible-versions.const';
 
 export abstract class BibleApi extends Bible {
   public abstract getPassage(osis: string, bibleVersion: BibleVersionCode): Promise<string>;
