+++ src/bible/online-bible/online-bible.class.ts
@@ -1,5 +1,5 @@
 import { Bible } from 'src/bible/bible.class';
-import { BibleVersionCode } from 'src/bible/versions/bible-versions.const';
+import { BibleVersionCode } from 'src/bible/models/bible-versions.const';
 
 export abstract class OnlineBible extends Bible {
   public abstract buildPassageLink(osis: string, bibleVersion: BibleVersionCode): string;
